+++ basyx.digitaltwinregistry/basyx.digitaltwinregistry.component/src/main/resources/banner.txt
@@ -1,9 +1,7 @@
-  ____           _____              
- |  _ \         / ____|             
- | |_) |  __ _ | (___   _   _ __  __
- |  _ <  / _` | \___ \ | | | |\ \/ /
- | |_) || (_| | ____) || |_| | >  < 
- |____/  \__,_||_____/  \__, |/_/\_\
-======================== __/ |======      
-AAS Environment         |___/                  
+  ____        ____               ____  _       _ _        _   _____          _         ____            _     _              
+ | __ )  __ _/ ___| _   ___  __ |  _ \(_) __ _(_) |_ __ _| | |_   _|_      _(_)_ __   |  _ \ ___  __ _(_)___| |_ _ __ _   _ 
+ |  _ \ / _` \___ \| | | \ \/ / | | | | |/ _` | | __/ _` | |   | | \ \ /\ / / | '_ \  | |_) / _ \/ _` | / __| __| '__| | | |
+ | |_) | (_| |___) | |_| |>  <  | |_| | | (_| | | || (_| | |   | |  \ V  V /| | | | | |  _ <  __/ (_| | \__ \ |_| |  | |_| |
+ |____/ \__,_|____/ \__, /_/\_\ |____/|_|\__, |_|\__\__,_|_|   |_|   \_/\_/ |_|_| |_| |_| \_\___|\__, |_|___/\__|_|   \__, |
+                    |___/                |___/                                                   |___/                |___/                 
 2.0.0-PREVIEW
